Fixes for conflicts
